Move merge_two_dicts to test utility file.
